move dataset ext trait to lance-index
